Fallback to resolving @imports as relative to resolve.modules
When resolving @imports, the WebpackFileManager assumes that they are
relative to the current file. However this is not neccessarily the case
as the less compiler allows specifying an `--include-paths` options (or
paths when rendering via js) and @imports may be specified relative to
any of these paths.

As a result less that compiles fine with the less compiler using the
--include-paths option cannot be compiled with the less-loader without
modifying all the @import statements to be relative to the file they are
in.

Removing the WebpackFileManager fixes the issue in terms of compilation
however this has the undesirable effect that dependencies are no longer
tracked correctly in Webpack. This means that modifying a file that is
imported will no longer trigger a rebuild when using the webpack `-w`
flag.

Therefore this modifies the WebpackFileManager so that if the initial
attempt to resolve the file fails, a second attempt is made with an
empty context and the original filename as specified in the less. This
means that by adding the include paths to the `resolve.modules` Webpack
configuration, the same effect is acheived as if using the less
--include-paths option whilst still tracking the dependencies correctly.